test: implement basic e2e tests for reporting endpoints
